+++ Runtime/Stages/VoxelShader.cs
@@ -1,6 +1,5 @@
 ﻿using UnityEngine;
 using UnityEngine.Rendering;
-using UnityEngine.Experimental.Rendering;
 
 public class VoxelShader : System.IDisposable {
   public ComputeShader compute {